# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API Store demo",
    "description": "Building a demo store portfolio",
    "contact": {
      "name": "Enderson Omana",
      "email": "omana.endersonj@gmail.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Please enter a valid token to test the requests below..."
    }
  },
  "paths": {
    "/users/": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get all users.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "User Information",
            "in": "body",
            "description": "",
            "schema": {
              "required": [
                "User"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "example": "example@mail.com"
                },
                "password": {
                  "type": "string",
                  "example": "ax12XSA&ads"
                },
                "role": {
                  "type": "string",
                  "example": "customer"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "User"
        ],
        "description": "This route permit update information about users, every field is optional",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "User Information",
            "in": "body",
            "description": "",
            "schema": {
              "required": [
                "User"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "example": "example@mail.com"
                },
                "role": {
                  "type": "string",
                  "example": "customer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}